Crate p3_symmetric

A framework for symmetric cryptography primitives.

Structs

CompressionFunctionFromHasher

Hash

A wrapper around an array digest, with a phantom type parameter to ensure that the digest is associated with a particular field.

MultiField32PaddingFreeSponge

A padding-free, overwrite-mode sponge function that operates natively over PF but accepts elements of F: PrimeField32.

PaddingFreeSponge

A padding-free, overwrite-mode sponge function.

SerializingHasher

Converts a hasher which can hash bytes, u32’s or u64’s into a hasher which can hash field elements.

TruncatedPermutation

Traits

CompressionFunction

An N-to-1 compression function.

CryptographicHasher

A generic trait for cryptographic hashers that consume an arbitrary sequence of input items and produce a fixed-size output.

CryptographicPermutation

A permutation thought to be cryptographically secure, in the sense that it is thought to be difficult to distinguish (in a nontrivial way) from a random permutation.

Permutation

A permutation in the mathematical sense.

PseudoCompressionFunction

An N-to-1 compression function collision-resistant in a hash tree setting.