Crate p3_symmetric

A framework for symmetric cryptography primitives.

Structs

CompressionFunctionFromHasher

Hash

A wrapper around an array digest, with a phantom type parameter to ensure that the digest is associated with a particular field.

MultiField32PaddingFreeSponge

A padding-free, overwrite-mode sponge function that operates natively over PF but accepts elements of F: PrimeField32.

PaddingFreeSponge

A padding-free, overwrite-mode sponge function.

SerializingHasher32

Maps input field elements to their 4-byte little-endian encodings, outputs [u8; 32].

SerializingHasher64

Maps input field elements to their 8-byte little-endian encodings, outputs [u8; 32].

TruncatedPermutation

Traits

CompressionFunction

An N-to-1 compression function.

CryptographicHasher

CryptographicPermutation

A permutation thought to be cryptographically secure, in the sense that it is thought to be difficult to distinguish (in a nontrivial way) from a random permutation.

Permutation

A permutation in the mathematical sense.

PseudoCompressionFunction

An n-to-1 compression function, like CompressionFunction, except that it need only be collision-resistant in a hash tree setting, where the preimage of a non-leaf node must consist of compression outputs.